Name change and code categorisation
Changed name to FireCracker
Started code categorisation
