+++ .gitignore
@@ -1,0 +1,5 @@
+bin/
+.classpath
+.project
+wordList.txt
+changelog.txt
