--- src/main/methods/CPU_optimized.java
@@ -1,11 +1,0 @@
-package main.methods;
-
-/*
- * Methods optimized to reduce CPU usage. 
- * These may be very RAM intensive.
- * 
- * */
-
-public class CPU_optimized {
-
-}
