--- src/main/methods/RAM_optimized.java
@@ -1,11 +1,0 @@
-package main.methods;
-
-/*
- * Methods optimized to reduce RAM usage 
- * These may be CPU and hard disk intensive.
- * 
- * */
-
-public class RAM_optimized {
-
-}
